Replace enzyme; remove enzyme configs
