+++ config/setupTests.js
@@ -1,8 +1,4 @@
 import React from 'react';
-import { configure, mount, render, shallow } from 'enzyme';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-
-configure({ adapter: new Adapter() });
 
 global.shallow = shallow;
 global.render = render;
