+++ jest.config.js
@@ -11,5 +11,4 @@ module.exports = {
     '/node_modules/(?!@redhat-cloud-services)',
     '/node_modules/(?!@patternfly)',
   ],
-  snapshotSerializers: ['enzyme-to-json/serializer'],
 };
