+++ package.json
@@ -15,7 +15,7 @@
     "start:beta": "BETA=true npm start",
     "start:beta:mock": "BETA=true MOCK=true npm start",
     "test:ct": "cypress run-ct --record false",
-    "test:jest": "jest --verbose --passWithNoTests --env=jsdom",
+    "test:jest": "jest --verbose --passWithNoTests",
     "test": "npm run test:ct && npm run test:jest",
     "translations": "npm-run-all translations:*",
     "translations:extract": "formatjs extract 'src/**/*.js*' --out-file lang/en.json --id-interpolation-pattern '[sha512:contenthash:base64:6]'",
@@ -54,14 +54,11 @@
     "@formatjs/cli": "4.2.27",
     "@redhat-cloud-services/eslint-config-redhat-cloud-services": "1.2.1",
     "@redhat-cloud-services/frontend-components-config": "^4.2.7",
-    "@wojtekmaj/enzyme-adapter-react-17": "0.6.2",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "27.0.6",
     "babel-plugin-dual-import": "1.2.1",
     "babel-plugin-transform-imports": "2.0.0",
     "cypress": "^8.4.0",
-    "enzyme": "3.11.0",
-    "enzyme-to-json": "3.6.2",
     "eslint": "7.30.0",
     "eslint-loader": "4.0.2",
     "eslint-plugin-cypress": "^2.12.1",
