+++ src/AppEntry.js
@@ -2,25 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { BrowserRouter as Router } from 'react-router-dom';
 
-import { IntlProvider } from '@redhat-cloud-services/frontend-components-translations/';
 import { getBaseName } from '@redhat-cloud-services/frontend-components-utilities/helpers/helpers';
 
 import App from './App';
-
-const translations = {
-  en: require('../compiled-lang/en.json'),
-};
-
-export const Intl = ({ children }) => (
-  <IntlProvider
-    locale={navigator.language.slice(0, 2)}
-    defaultLocale="en"
-    messages={translations[navigator.language.slice(0, 2)]}
-    onError={console.error}
-  >
-    {children}
-  </IntlProvider>
-);
+import { Intl } from './Utilities/intlHelper';
 
 const AppEntry = ({ useLogger }) => (
   <Intl>
@@ -29,10 +14,6 @@ const AppEntry = ({ useLogger }) => (
     </Router>
   </Intl>
 );
-
-Intl.propTypes = {
-  children: PropTypes.element,
-};
 
 AppEntry.propTypes = {
   useLogger: PropTypes.bool,
